multiprecision: remove detail/type_traits and clean up
