+++ crypto3/libs/multiprecision/include/nil/crypto3/multiprecision/detail/big_mod/ops/pow.hpp
@@ -19,11 +19,11 @@
 #include "nil/crypto3/multiprecision/detail/big_mod/big_mod_impl.hpp"
 #include "nil/crypto3/multiprecision/detail/big_mod/ops/inverse.hpp"
 #include "nil/crypto3/multiprecision/detail/integer_utils.hpp"
-#include "nil/crypto3/multiprecision/detail/type_traits.hpp"
 
 namespace nil::crypto3::multiprecision {
     template<typename big_mod_t, typename T,
-             std::enable_if_t<detail::is_big_mod_v<big_mod_t> && detail::is_unsigned_integer_v<T>,
+             std::enable_if_t<detail::is_big_mod_v<big_mod_t> && detail::is_integral_v<T> &&
+                                  !std::numeric_limits<T>::is_signed,
                               int> = 0>
     constexpr big_mod_t pow(const big_mod_t &b, const T &e) {
         big_mod_t result(b.ops_storage());
@@ -32,7 +32,7 @@ namespace nil::crypto3::multiprecision {
     }
 
     template<typename big_mod_t, typename T,
-             std::enable_if_t<detail::is_big_mod_v<big_mod_t> && detail::is_integer_v<T> &&
+             std::enable_if_t<detail::is_big_mod_v<big_mod_t> && detail::is_integral_v<T> &&
                                   std::numeric_limits<T>::is_signed,
                               int> = 0>
     constexpr big_mod_t pow(const big_mod_t &b, const T &e) {
