+++ crypto3/libs/multiprecision/include/nil/crypto3/multiprecision/detail/integer_ops_base.hpp
@@ -4,7 +4,6 @@
 #include <type_traits>
 
 #include "nil/crypto3/multiprecision/detail/big_uint/big_uint_impl.hpp"
-#include "nil/crypto3/multiprecision/detail/type_traits.hpp"
 
 namespace nil::crypto3::multiprecision {
     template<typename T, std::enable_if_t<std::is_integral_v<T>, int> = 0>
